+++ temp/LectureDonneesSimulateur.java
@@ -137,7 +137,7 @@ public class LectureDonneesSimulateur {
         try {
             int lig = scanner.nextInt();
             int col = scanner.nextInt();
-            int intensite = scanner.nextInt();
+            double intensite = scanner.nextInt();
             if (intensite <= 0) {
                 throw new DataFormatException("incendie " + i
                         + "nb litres pour eteindre doit etre > 0");
